Conflicts:
	src/lib/modules/IntEqModule/IntEqModule.tpp

+++ src/lib/modules/GBModule/GroebnerModule.h
@@ -91,7 +91,7 @@ protected:
     /// The rewrite rules for the variables
     groebner::RewriteRules mRewriteRules;
 
-    std::map<unsigned, carl::Variable> mAdditionalVarMap;
+    std::map<size_t, carl::Variable> mAdditionalVarMap;
     
     /** A workaround to associate equalities in the passed formula originating from the gb
      * (in contrast to those which originate from simplified formulae)
